IDL-60 Use vault integration if Vault client is initialised successfully in the backend
